backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.35.8
pydantic==2.8.0
pydantic_core==2.20.0
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
